requirements.txt
glfw~=2.1.0
numpy~=1.20.3
PyOpenGL~=3.1.5